Refactor: Reorganize namespaces for better modularity
Reorganized `Talabat.Domain.Contracts.Presistence` into sub-namespaces:
- Moved `ISpecifications` to `DbIntializers.Specification`.
- Moved `IDbIntializer`, `IStoreDbIntializer`, and `IStoreIdentityInializer` to `DbIntializers`.
- Moved `IGenericRepository` to `GenericRepository`.
- Moved `IUnitOfWork` to `UnitOfWork`.

Updated classes (`BaseSpecifications`, `DependencyInjection`, `GenericRepository`, `SpecificationsEvaluator`, `UnitOfWork`, `StoreDbInitializer`, etc.) to use the new namespaces.

Removed old interface definitions and reintroduced them in their respective new namespaces.

This refactoring improves code organization and modularity.
