+++ Talabat.APIs/Extentions/InitializerExtentions.cs
@@ -1,4 +1,4 @@
-﻿using Talabat.Domain.Contracts.Presistence;
+﻿using Talabat.Domain.Contracts.Presistence.DbIntializers;
 using Talabat.Infrastructure.Persistence;
 
 namespace Talabat.APIs.Extentions
