+++ Talabat.Domain/Contracts/Presistence/DbIntializers/IStoreDbIntializer.cs
@@ -1,4 +1,4 @@
-﻿namespace Talabat.Domain.Contracts.Presistence
+﻿namespace Talabat.Domain.Contracts.Presistence.DbIntializers
 {
     public interface IStoreDbIntializer:IDbIntializer
     {
