+++ Talabat.Infrastructure.Persistence/DependencyInjection.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Talabat.Domain.Contracts.Presistence.DbIntializers;
+using Talabat.Domain.Contracts.Presistence.UnitOfWork;
 using Talabat.Infrastructure.Persistence._Data;
 using Talabat.Infrastructure.Persistence._Identity;
 namespace Talabat.Infrastructure.Persistence
