+++ Talabat.Infrastructure.Persistence/_Data/StoreDbInitializer.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text.Json;
+using Talabat.Domain.Contracts.Presistence.DbIntializers;
 using Talabat.Domain.Entities.Products;
 
 namespace Talabat.Infrastructure.Persistence._Data
